requirements.txt
azure-functions
azure-storage-blob
fastavro
pyodbc
